Rename utility folder (capitalize first letter)
