+++ src/app/components/features/profile-page/edit-password/edit-password.component.ts
@@ -1,5 +1,5 @@
 import {Component, computed, inject, OnInit, signal, Signal } from '@angular/core';
-import {PasswordValidator, UpdatePasswordErrors} from '../../../../core/Utility/Validation';
+import {PasswordValidator, UpdatePasswordErrors} from '../../../../core/utility/Validation';
 import {FormControl, FormGroup, ReactiveFormsModule, Validators} from '@angular/forms';
 import {UserService} from '../../../../core/services/user-service/user.service';
 import {ScaledText} from '../../../shared/scaled-text/scaled-text';
